app/requirements.txt
asgiref==3.7.2
Django==4.2
django-localflavor==4.0
python-decouple==3.8
python-stdnum==1.19
sqlparse==0.4.4
typing_extensions==4.9.0
